backend/requirements.txt
# Core Flask Web Framework
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# Machine Learning & Deep Learning
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
scikit-learn==1.3.0

# Medical Imaging
nibabel==5.1.0
SimpleITK==2.2.1
pydicom==2.4.2

# Image Processing  
opencv-python==4.8.0.74
Pillow==10.0.0

# Data Handling
pandas==2.0.3

# Utilities
python-dotenv==1.0.0
pathlib
flask==2.3.0
flask-cors==4.0.0
torch==2.1.0+cpu
torchvision==0.16.0+cpu
Pillow==10.0.0
gunicorn==21.2.0
